Use FreeTypeGX
Specifically a version I created to work alongside GRRLIB: https://github.com/HTV04/freetypegx-grrlib

By proxy, this also means the Metaphrasis library has been added. Could it be useful? Maybe...

As of writing, this commit uses a custom version of GRRLIB with 2D transformation functionality. I created a PR for this functionality, so hopefully this won't remain the case.